+++ src/bluesearch/entrypoint/database/parent.py
@@ -15,6 +15,7 @@ from bluesearch.entrypoint.database import (
     parse,
     parse_mesh_rdf,
     topic_extract,
+    topic_filter,
 )
 
 Cmd = namedtuple("Cmd", ["help", "init_parser", "run"])
@@ -75,6 +76,11 @@ def main(argv: Sequence[str] | None = None) -> int:
             help="Extract topic of article(s).",
             init_parser=topic_extract.init_parser,
             run=topic_extract.run,
+        ),
+        "topic-filter": Cmd(
+            help="Filter articles containing relevant topics.",
+            init_parser=topic_filter.init_parser,
+            run=topic_filter.run,
         ),
         "parse-mesh-rdf": Cmd(
             help="Parse a MeSH RDF file to extract the topic tree structure.",